some test updates
