+++ packages/component/src/Activity/private/feedbackActivity.util.ts
@@ -1,10 +1,10 @@
 import { type WebChatActivity } from 'botframework-webchat-core';
-import { object, safeParse, string } from 'valibot';
+import { literal, object, safeParse, string } from 'valibot';
 
 const activityWithFeedbackLoopSchema = object({
   channelData: object({
     feedbackLoop: object({
-      type: string(),
+      type: literal('default'),
       disclaimer: string()
     })
   })
@@ -15,7 +15,7 @@ type FeedbackActivity = WebChatActivity & {
 };
 
 export const hasFeedbackLoop = (activity: WebChatActivity): activity is FeedbackActivity =>
-  safeParse(activityWithFeedbackLoopSchema, activity.channelData).success;
+  safeParse(activityWithFeedbackLoopSchema, activity).success;
 
 export const getDisclaimer = (activity: WebChatActivity): string | undefined =>
   hasFeedbackLoop(activity) ? activity.channelData.feedbackLoop.disclaimer : undefined;
